refactor: separate data source provider
